fix(gateway): retain publication recovery authority
